empty print().py
We've added one empty print() function invocation. We call it empty because we haven't delivered any arguments to the function.

You can see it in the editor window. Run the code.

What happens?

If everything goes right, you should see something like this:

INPUT :-

![p8](https://user-images.githubusercontent.com/16610336/96480740-4476f200-1258-11eb-9697-e4efe840c6c6.png)


OUTPUT :-
 
![p9](https://user-images.githubusercontent.com/16610336/96480746-450f8880-1258-11eb-9417-4fe6c944f6cd.png)

As you can see, the empty print() invocation is not as empty as you may have expected - it does output an empty line, or (this interpretation is also correct) its output is just a newline.

But this is not the only way to produce a newline in the output console.