escape and newline character.py
We have done  two very subtle changes - we've inserted a strange pair of characters inside the rhyme. They look like this: ' \n.'

While you can see two characters, Python sees one.

The backslash (\) has a very special meaning when used inside strings - this is called the escape character.

The word escape should be understood specifically - it means that the series of characters in the string escapes for the moment (a very short moment) to introduce a special inclusion.

In other words, the backslash doesn't mean anything in itself, but is only a kind of announcement, that the next character after the backslash has a different meaning too.

The letter n placed after the backslash comes from the word newline.
Both the backslash and the n form a special symbol named a newline character, which urges the console to start a new output line.

Run the code. Your console should now look like this:

INPUT :- 


![P10](https://user-images.githubusercontent.com/16610336/96485836-4e99f000-125a-11eb-833d-a5f99ceece0f.png)

OUTPUT :-

![P11](https://user-images.githubusercontent.com/16610336/96485852-4fcb1d00-125a-11eb-9e4b-bd3695f191ea.png)

As you can see, two newlines appear in the nursery rhyme, in the places where the \n have been used.