keyword arguments.py
Python offers another mechanism for the passing of arguments, which can be helpful when you want to convince the print() function to change its behavior a bit.

The mechanism is called keyword arguments. The name stems from the fact that the meaning of these arguments is taken not from its location (position) but from the special word (keyword) used to identify them.

The print() function has two keyword arguments that you can use for your purposes. The first of them is named "end".

In order to use it, it is necessary to know some rules:

    a keyword argument consists of three elements: a keyword identifying the argument (end here); an equal sign (=); and a value assigned to that argument;
    any keyword arguments have to be put after the last positional argument (this is very important)


In my example, we have made use of the end keyword argument, and set it to a string containing one space.

Run the code to see how it works.

INPUT :- 

![P12](https://user-images.githubusercontent.com/16610336/96572824-f3fca480-12ea-11eb-8524-569b7882578f.png)

OUTPUT :-

![P13](https://user-images.githubusercontent.com/16610336/96572831-f52dd180-12ea-11eb-9d58-12e5c60256ca.png)

As you can see, the end keyword argument determines the characters the print() function sends to the output once it reaches the end of its positional arguments.